server: move the server to its own folder
We want a clean structure with the client and server so we can
articulate them all via Docker (Compose).
