--- index.ts
@@ -1,3 +1,0 @@
-import { runServer } from './server'
-
-runServer()
