+++ server/index.ts
@@ -1,3 +1,22 @@
-import { runServer } from './runServer'
+import Express from "express"
+import cors from 'cors'
+import { buildRouter } from "./router";
+import { client } from "./database";
 
-export { runServer }
+const PORT = 3001;
+
+const app = Express();
+
+const router = buildRouter();
+
+app.use(cors());
+
+app.use(Express.json())
+
+app.use("/api", router)
+
+app.listen(PORT, async () => {
+  console.log(`Connecting to the database...`)
+  await client.connect();
+  console.log(`Server is running on port ${PORT}`)
+})
