+++ server/package.json
@@ -4,16 +4,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "buildServer": "tsc",
-    "buildClient": "cd client && npm run build",
-    "db:up": "docker-compose up",
-    "db:down": "docker-compose down",
-    "runServer": "npm run buildServer && node dist",
-    "runClient": "cd client && npm run start",
-    "start": "concurrently \"npm run runServer\" \"npm run runClient\" ",
-    "build": "concurrently \"npm run buildServer\" \"npm run buildClient\" ",
-    "runProd": "concurrently \"node dist/serveClient.js\" \"node dist\"",
-    "runFixtures": "tsc && node dist/server/scripts/runFixtures.js",
+    "build": "tsc",
+    "start": "node dist",
+    "runFixtures": "npm run build && node dist/server/scripts/runFixtures.js",
+    "knex": "knex",
     "migrate:up": "knex migrate:up"
   },
   "author": "",
