--- server/runServer.ts
@@ -1,26 +1,0 @@
-import Express from "express"
-import cors from 'cors'
-import { buildRouter } from "./router";
-import { client } from "./database";
-
-export { runServer }
-
-const PORT = 3001;
-
-function runServer() {
-    const app = Express();
-
-    const router = buildRouter();
-
-    app.use(cors());
-
-    app.use(Express.json())
-
-    app.use("/api", router)
-
-    app.listen(PORT, async () => {
-        console.log(`Connecting the database...`)
-        await client.connect();
-        console.log(`Server is running on port ${PORT}`)
-    })
-}
